Create set-membership.py
Due to a bug with Jira Cloud import (https://jira.atlassian.com/browse/JRACLOUD-67859), there is a need to contact Atlassian support to restore all users, group and membership. This script allow to do part of this job by restoring membership in a target Jira based on a DB dump.